inference-service/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
python-multipart>=0.0.12,<1.0
pillow>=10.0,<12.0
torch>=2.5,<3.0
torchvision>=0.20,<1.0
